Support for Canada Label - WELL-2968, WELL-2959, WELL-2971
